@@ -4,7 +4,7 @@ publish_to: "none"
 version: "1.0.0+25"
 
 environment:
-  sdk: ">=2.19.0 <4.0.0"
+  sdk: ">=3.6.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -13,6 +13,7 @@ dependencies:
   json_dynamic_widget_plugin_rive:
     path: "../"
   logging: "^1.3.0"
+  rive: "^0.14.0"
 
 dev_dependencies:
   flutter_lints: "^6.0.0"
